# docs/openapi.json
{
  "components": {
    "schemas": {
      "CashEventResource": {
        "properties": {
          "amountPoisha": {
            "title": "Amountpoisha",
            "type": "integer"
          },
          "countedCashPoisha": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Countedcashpoisha"
          },
          "eventId": {
            "title": "Eventid",
            "type": "string"
          },
          "expectedAfterPoisha": {
            "title": "Expectedafterpoisha",
            "type": "integer"
          },
          "expectedBeforePoisha": {
            "title": "Expectedbeforepoisha",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "CASH_IN",
              "CASH_OUT",
              "RECONCILIATION"
            ],
            "title": "Kind",
            "type": "string"
          },
          "observedAt": {
            "format": "date-time",
            "title": "Observedat",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "recordedAt": {
            "format": "date-time",
            "title": "Recordedat",
            "type": "string"
          },
          "sequenceNumber": {
            "title": "Sequencenumber",
            "type": "integer"
          },
          "source": {
            "enum": [
              "SIMULATOR",
              "DEVICE",
              "USER"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "sequenceNumber",
          "kind",
          "amountPoisha",
          "expectedBeforePoisha",
          "expectedAfterPoisha",
          "countedCashPoisha",
          "source",
          "reason",
          "observedAt",
          "recordedAt"
        ],
        "title": "CashEventResource",
        "type": "object"
      },
      "CashMutationResource": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/CashEventResource"
          },
          "wallet": {
            "$ref": "#/components/schemas/SmartWalletResource"
          }
        },
        "required": [
          "event",
          "wallet"
        ],
        "title": "CashMutationResource",
        "type": "object"
      },
      "CashObservationBody": {
        "additionalProperties": false,
        "properties": {
          "amountPoisha": {
            "exclusiveMinimum": 0.0,
            "title": "Amountpoisha",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "CASH_IN",
              "CASH_OUT"
            ],
            "title": "Kind",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 140,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "kind",
          "amountPoisha"
        ],
        "title": "CashObservationBody",
        "type": "object"
      },
      "ConnectionBody": {
        "additionalProperties": false,
        "properties": {
          "connected": {
            "title": "Connected",
            "type": "boolean"
          }
        },
        "required": [
          "connected"
        ],
        "title": "ConnectionBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReconciliationBody": {
        "additionalProperties": false,
        "properties": {
          "countedCashPoisha": {
            "minimum": 0.0,
            "title": "Countedcashpoisha",
            "type": "integer"
          },
          "reason": {
            "maxLength": 140,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "countedCashPoisha",
          "reason"
        ],
        "title": "ReconciliationBody",
        "type": "object"
      },
...
      "SmartWalletResource": {
        "properties": {
          "activity": {
            "items": {
              "$ref": "#/components/schemas/CashEventResource"
            },
            "title": "Activity",
            "type": "array"
          },
          "connectionStatus": {
            "enum": [
              "CONNECTED",
              "DISCONNECTED"
            ],
            "title": "Connectionstatus",
            "type": "string"
          },
          "expectedCashPoisha": {
            "title": "Expectedcashpoisha",
            "type": "integer"
          },
          "inventoryDifferencePoisha": {
            "title": "Inventorydifferencepoisha",
            "type": "integer"
          },
          "lastSequence": {
            "title": "Lastsequence",
            "type": "integer"
          },
          "lastSyncedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastsyncedat"
          },
          "walletId": {
            "title": "Walletid",
            "type": "string"
          }
        },
        "required": [
          "walletId",
          "connectionStatus",
          "expectedCashPoisha",
          "lastSequence",
          "lastSyncedAt",
          "inventoryDifferencePoisha",
          "activity"
        ],
        "title": "SmartWalletResource",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Money Movement",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/smart-wallet": {
      "get": {
        "operationId": "get_smart_wallet_api_v1_smart_wallet_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartWalletResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Smart Wallet",
        "tags": [
          "smart-wallet"
        ]
      }
    },
    "/api/v1/smart-wallet/connection": {
      "post": {
        "operationId": "set_connection_api_v1_smart_wallet_connection_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartWalletResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Connection",
        "tags": [
          "smart-wallet"
        ]
      }
    },
    "/api/v1/smart-wallet/events": {
      "post": {
        "operationId": "record_cash_event_api_v1_smart_wallet_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashObservationBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashMutationResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Cash Event",
        "tags": [
          "smart-wallet"
        ]
      }
    },
    "/api/v1/smart-wallet/reconciliations": {
      "post": {
        "operationId": "reconcile_cash_api_v1_smart_wallet_reconciliations_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconciliationBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashMutationResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile Cash",
        "tags": [
          "smart-wallet"
        ]
      }
    },
...
  }
}